Separate apps in to components
