+++ ui/src/components/InputDeviceHostname.vue
@@ -1,73 +1,4 @@
 <template>
-  <q-bar v-if="$q.platform.is.electron" class="bg-accent q-electron-drag" dense>
-    <div class="row justify-between content-center full-width no-wrap">
-      <div class="col-4 no-wrap">
-        <q-btn
-          dense
-          flat
-          round
-          icon="lens"
-          size="8.5px"
-          color="red"
-          @click="electronCloseApp"
-        />
-        <q-btn
-          dense
-          flat
-          round
-          icon="lens"
-          size="8.5px"
-          color="yellow"
-          @click="electronMinimize"
-        />
-        <q-btn
-          dense
-          flat
-          round
-          icon="lens"
-          size="8.5px"
-          color="green"
-          @click="electronToggleMaximize"
-        />
-      </div>
-      <div class="no-wrap">
-        <q-avatar
-          v-if="$q.platform.is.electron"
-          square
-          size="xs"
-          class="q-mr-sm"
-        >
-          <img :src="qHeaderStyle.logo_coloured" /> </q-avatar
-        >{{ $t('title') }}
-      </div>
-      <div class="col-4"></div>
-    </div>
-  </q-bar>
-  <q-slide-transition>
-    <q-banner
-      v-if="showInstallBanner"
-      dense
-      inline-actions
-      class="text-white bg-primary"
-    >
-      {{ $t('system.pwa.install_notification') }}
-      <template #action>
-        <q-btn
-          flat
-          color="white"
-          :label="$t('general.close')"
-          @click="installApp(false)"
-        />
-        <q-btn
-          flat
-          color="white"
-          :label="$t('system.pwa.install')"
-          @click="installApp(true)"
-        />
-      </template>
-    </q-banner>
-  </q-slide-transition>
-
   <q-toolbar class="bg-accent">
     <q-avatar v-if="quasarMode === 'pwa'" square size="sm">
       <img :src="qHeaderStyle.logo_coloured" />
@@ -182,36 +113,12 @@
 import DeviceInfo from 'components/SystemDeviceInfo.vue'
 import { qHeaderStyle } from 'src/config/qStyles'
 import { axiosSettings } from 'stores/system'
-import { defineComponent, onMounted, ref } from 'vue'
-import { useQuasar, QInput } from 'quasar'
-
-interface BeforeInstallPromptEvent extends Event {
-  readonly platforms: string[]
-  readonly userChoice: Promise<{
-    outcome: 'accepted' | 'dismissed'
-    platform: string
-  }>
-  prompt(): Promise<void>
-}
-
-declare global {
-  interface WindowEventMap {
-    beforeinstallprompt: BeforeInstallPromptEvent
-  }
-}
+import { defineComponent, ref } from 'vue'
+import { QInput } from 'quasar'
 
 interface tabIndex {
   title: string
 }
-
-declare const window: Window &
-  typeof globalThis & {
-    myWindowAPI: {
-      close: () => void
-      minimize: () => void
-      toggleMaximize: () => void
-    }
-  }
 
 export default defineComponent({
   name: 'LayoutsCaptivePortal',
@@ -221,28 +128,10 @@ export default defineComponent({
   setup() {
     const axiosBaseUrl = axiosSettings()
     const currentTab = ref('welcome')
-    const showInstallBanner = ref(false)
     const quasarMode = ref(process.env.MODE)
     const reqHostname = ref('')
-    const $q = useQuasar()
     const tabs = ref<tabIndex[]>([])
     const validateHostInput = ref<QInput>()
-
-    let deferredPrompt: BeforeInstallPromptEvent
-
-    onMounted(() => {
-      if (
-        process.env.MODE === 'pwa' &&
-        $q.localStorage.getItem('dismissedInstall') !== true
-      ) {
-        window.addEventListener('beforeinstallprompt', (event) => {
-          event.preventDefault()
-          deferredPrompt = event
-
-          showInstallBanner.value = true
-        })
-      }
-    })
 
     function addHttp(url: string) {
       if (!/^(?:f|ht)tps?:\/\//.test(url)) {
@@ -251,39 +140,9 @@ export default defineComponent({
       return url
     }
 
-    // Electron toolbar controls
     function closeTab(index: number) {
       currentTab.value = 'welcome'
       tabs.value.splice(index, 1)
-    }
-
-    function electronCloseApp() {
-      window.myWindowAPI.close()
-    }
-
-    function electronMinimize() {
-      window.myWindowAPI.minimize()
-    }
-
-    function electronToggleMaximize() {
-      window.myWindowAPI.toggleMaximize()
-    }
-
-    function installApp(install: boolean) {
-      if (!install) {
-        showInstallBanner.value = false
-        $q.localStorage.set('dismissedInstall', true)
-        return
-      }
-
-      try {
-        void deferredPrompt.prompt()
-        showInstallBanner.value = false
-        $q.localStorage.set('dismissedInstall', true)
-      } catch (error) {
-        console.error(error)
-        showInstallBanner.value = false
-      }
     }
 
     function setAxios(currentTabSetting: string) {
@@ -306,17 +165,12 @@ export default defineComponent({
 
     return {
       closeTab,
-      electronMinimize,
-      electronToggleMaximize,
-      electronCloseApp,
       currentTab,
-      installApp,
       qHeaderStyle,
       quasarMode,
       reqHostname,
       setAxios,
       setHostname,
-      showInstallBanner,
       tabs,
       validateHostInput
     }
