+++ ui/src/components/MainLayoutMenuItems.vue
