+++ ui/src/layouts/MainLayout.vue
@@ -130,7 +130,7 @@ import { AxiosResponse } from 'axios'
 import { loadLanguageAsync } from 'boot/i18n'
 import Reboot from 'components/SystemReboot.vue'
 import Shutdown from 'components/SystemShutdown.vue'
-import MenuItems from 'layouts/MenuItems.vue'
+import MenuItems from 'components/MainLayoutMenuItems.vue'
 import { supervisor } from 'src/api/supervisor'
 import localeOptions from 'src/config/localeOptions'
 import menuList from 'src/config/menuList'
